Rename files and legend env var, DRY up styling
