+++ src/styles/legendAchromatic.js
