+++ src/styles/legendColored.js
